requirements.txt
azure-search-documents
python-dotenv
pydantic
openai
pytest
requests
